feat: 🎸 add loading for all pages
