+++ src/app/(dumper)/@contents/page.tsx
@@ -1,21 +1,13 @@
-import { CardStackSkeleton } from "@/components/stack/card-stack-skeleton";
-import { Separator } from "@/components/ui/separator";
-import { hasDumperPostPermission } from "@/features/auth/utils/role";
-import { AddContentsForm } from "@/features/contents/components/add-contents-form";
-import { ContentsStack } from "@/features/contents/components/contents-stack";
+import Loading from "@/components/loading";
 import { Suspense } from "react";
+import { SuspensePage } from "./_page";
 
 export const dynamic = "force-dynamic";
 
 export default async function Page() {
-	const hasPostPermission = await hasDumperPostPermission();
 	return (
-		<>
-			{hasPostPermission && <AddContentsForm />}
-			<Separator className="h-px bg-gradient-to-r from-primary to-primary-grad" />
-			<Suspense fallback={<CardStackSkeleton />}>
-				<ContentsStack />
-			</Suspense>
-		</>
+		<Suspense fallback={<Loading />}>
+			<SuspensePage />
+		</Suspense>
 	);
 }
